adds pagimation module
