+++ index.html
@@ -27,11 +27,7 @@
     <div class="container">
         <nav id="pagination"  aria-label="Page navigation example">
             <ul class="pagination  d-flex flex-wrap">
-             <!--  <li class="page-item"><a class="page-link" href="#">Previous</a></li>
-              <li class="page-item"><a class="page-link" href="#">1</a></li>
-              <li class="page-item"><a class="page-link" href="#">2</a></li>
-              <li class="page-item"><a class="page-link" href="#">3</a></li>
-              <li class="page-item"><a class="page-link" href="#">Next</a></li> -->
+            
             </ul>
           </nav>
