+++ main.js
@@ -2,14 +2,17 @@ import { getPokemons, getPokemonById } from "./services/services.js";
 
 import {
   createPokemonsCards,
-  createPagination,
   clearPreviousElements,
   setDetailModal,
   clearCards,
   showLoader,
-  handlePagination,
-  handleSelectedPage
 } from "./ui/ui.js";
+
+import {
+  createPagination,
+  handlePagination,
+  handleSelectedPage,
+} from "./pagination/pagination.js";
 
 async function initialize(offset = 0) {
   let pages;
@@ -26,15 +29,17 @@ async function initialize(offset = 0) {
 
   pages = Math.ceil(totalPokemons / pokemonsPerPage);
 
-  createPagination(pages, updatePokemonsCards, handlePagination,handleSelectedPage);
+  createPagination(
+    pages,
+    updatePokemonsCards,
+    handlePagination,
+    handleSelectedPage
+  );
 
   createPokemonsCards(pokemons, handlePokemonDetails);
 }
 
-async function updatePokemonsCards(offset, e) {
-  // console.log(e.target);
-  //stops anchor in pagination of changing URL #
-  // e.preventDefault();
+async function updatePokemonsCards(offset) {
   let pokemonsData = await getPokemons(offset);
   let pokemons = pokemonsData.results;
